Major rewrite of MateFinder, uses TTs now, PV extracted from TT so we don't need the PV arrays anymore. Bugfix where unmove of promotion was bugged and lead to a crash. Use AttackBoards for inCheck method calls for check generation. Introduce correspondence variable to enable or disable lichess CC time control settings as well as the locking mechanisms.

+++ KK_engine/src/Main/MultiThreading/AggressiveMatefinderTT.java
@@ -7,7 +7,7 @@ public class AggressiveMatefinderTT extends AbstractMatefinderTT {
     /**
      * @param sizeInByte power of 2.
      */
-    public AggressiveMatefinderTT(int sizeInByte) {
+    public AggressiveMatefinderTT(long sizeInByte) {
         super(sizeInByte);
     }
 }
